# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Slack Alarm Microservice API
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alarms:
    x-swagger-router-controller: alarms
    post:
      tags:
        - alarm
      summary: create alarm
      description: accepts a new alarm and returns same
      operationId: createAlarm
      parameters:
        - in: body
          name: alarm
          description: the alarm details to create
          required: true
          schema:
            $ref: '#/definitions/Alarm'
      responses:
        '201':
          description: success, new alarm created
          schema:
            $ref: '#/definitions/Alarm'
        default:
          description: error, failed
          schema:
            $ref: '#/definitions/ErrorResponse'
    get:
      tags:
        - alarms
      summary: get all alarms
      description: returns a list of all alarms
      operationId: getAllAlarms
      responses:
        '200':
          description: success, got list of alarms
          schema:
            type: array
            items:
              $ref: '#/definitions/Alarm'
        default:
          description: error, failed to get alarms list
          schema:
            $ref: '#/definitions/ErrorResponse'
  /alarms/{id}:
    x-swagger-router-controller: alarms
    get:
      tags:
        - alarm
      summary: get alarm details
      description: get details of a single alarm
      operationId: getAlarmDetails
      parameters:
        - name: id
          in: path
          description: get details of a particular alarm
          required: true
          type: string
      responses:
        '200':
          description: success, got alarm
          schema:
            $ref: '#/definitions/Alarm'
        default:
          description: error, failed to get alarm details
          schema:
            $ref: '#/definitions/ErrorResponse'
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  Alarm:
    type: object
    required:
      - id
      - name
      - alertAt
    properties:
      id:
        type: string
        pattern: '^[\da-fA-F]{8}-[\da-fA-F]{4}-[\da-fA-F]{4}-[\da-fA-F]{4}-[\da-fA-F]{12}'
      name:
        type: string
      alertAt:
        type: string
        format: date-time
#        pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(|\.[0-9]{1,3})Z$'
  ErrorResponse:
    type: object
    required:
      - message
    properties:
      message:
        type: string